# spec/src/main/resources/swagger/openapi.yml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: Localhost
    url: http://localhost:8080/user-app/v1
info:
  description: This is a user API
  version: "v1"
  title: User App
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: admins
    description: Secured Admin-only calls
  - name: developers
    description: Operations available to regular developers
paths:
  /users:
    get:
      tags:
        - developers
      summary: searches users
      operationId: searchUser
      description: |
        By passing in the appropriate options, you can search for
        available user in the system
      parameters:
        - in: query
          name: searchString
          description: pass an optional search string for looking up inventory
          required: false
          schema:
            type: string
        - in: query
          name: skip
          description: number of records to skip for pagination
          schema:
            type: integer
            format: int32
            minimum: 0
        - in: query
          name: limit
          description: maximum number of records to return
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 50
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: bad input parameter
    post:
      tags:
        - admins
      summary: adds an user
      operationId: addUser
      description: Adds an item to the system
      responses:
        '201':
          description: item created
        '400':
          description: 'invalid input, object invalid'
        '409':
          description: an existing item already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: Inventory item to add
components:
  schemas:
    User:
      type: object
      required:
        - id
        - name
        - place
      properties:
        id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        name:
          type: string
          example: User name
        place:
          type: string
          example: User place